eval/requirements.txt
coloredlogs==15.0.1
cuda-python==11.8.3
cudf-cu11==24.8.2
cupy-cuda11x==13.3.0
fastapi==0.112.2
flatbuffers==24.3.25
humanfriendly==10.0
milvus-model==0.2.7
minio==7.2.8
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvtx==0.2.10
onnxruntime==1.19.2
pandas==2.2.2
pymilvus==2.4.5
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
tqdm==4.66.5
torch==1.13.1
torchvision==0.14.1
transformers==4.44.1
